Fix CI 🤞
Update the SauceLabs configuration for updates to the their service.
The have now removed ALL versions of Opera and don't appear to support
any iOS versions before 9.3.

There is a transient dependency on sqlite3 which does not build on
Travis because the sqlite3 node module does not yet support node 10
and the Travis build runs on the latest "stable" node, which is
currently 10. I specified node 8 in .travis.yml.